Move demo site content definitions around
